refactor(LikesException) : 예외 및 예외 메시지 코드 패키지 이동
